chore/prepare cypress setup
