+++ development/setup.js
@@ -1,51 +1,11 @@
-const fs = require("fs");
-const path = require("path");
-const os = require("os");
 const gitaApi = require("./utils/gitea-api.js");
 const waitFor = require("./utils/wait-for.js");
 const runCommand = require("./utils/run-command.js");
-
-const randomPass = () =>
-  [
-    Math.random().toString(36).substring(2, 5),
-    Math.random().toString(36).substring(2, 5),
-  ].join("DIG@");
-
-const defaultEnvVars = {
-  DEVELOP_BACKEND: 0,
-  DEVELOP_DASHBOARD: 0,
-  DEVELOP_APP_DEVELOPMENT: 0,
-  DEVELOP_PREVIEW: 0,
-  GITEA_ADMIN_USER: "localgiteaadmin",
-  GITEA_ADMIN_PASS: randomPass(),
-  CYPRESS_TEST_APP: "autodeploy-v3",
-  GITEA_ORG_USER: "ttd",
-};
-
-const ensureDotEnv = () => {
-  const dotenvLocations = path.resolve(__dirname, "..", ".env");
-  const envData = { ...defaultEnvVars };
-  const existingData = fs.existsSync(dotenvLocations)
-    ? fs.readFileSync(dotenvLocations, "utf-8").split(os.EOL)
-    : [];
-  existingData.forEach((line) => {
-    const trimmedLine = line.trim();
-    if (trimmedLine.length > 0 && trimmedLine[0] !== "#") {
-      const [key, value] = trimmedLine.split("=");
-      envData[key] = value;
-    }
-  });
-  const newEnv = [];
-  Object.keys(envData).forEach((key) =>
-    newEnv.push([key, envData[key]].join("="))
-  );
-  fs.writeFileSync(dotenvLocations, newEnv.join(os.EOL), "utf-8");
-  console.log("Done ensuring .env variables at:", dotenvLocations);
-  return envData;
-};
+const ensureDotEnv = require("./utils/ensure-dot-env.js");
 
 const startingDockerCompose = () =>
   runCommand("docker compose up -d --remove-orphans");
+
 const createAdminUser = (env) =>
   runCommand(
     [
